+++ api/templates/swagger/oauth2-redirect.html
